// src/app/generated/api/model/itemFields.ts
/**
 * Listok API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemStatus } from './itemStatus';
import { Product } from './product';


/**
 * Fields required to create or update an Item.
 */
export interface ItemFields { 
    product: Product;
    /**
     * Quantity of the Item to be purchased.
     */
    count: number;
    /**
     * Optional color associated with the Item.
     */
    groupColor?: string;
    status: ItemStatus;
}